Move zero-effect to it's own package.
